Removed the parse_args stuff, implemented centralized argument parsing, partially implemented a variable args solution, partially converted MooInteger into a type capable of handling both ints and floats, so that we don't have to do math with 2 different types all the time, and also added some functions to array and basic.

+++ src/actions/builtin/room.cpp
@@ -104,7 +104,7 @@ static int room_go(MooCodeFrame *frame, MooObjectHash *env, MooArgs *args)
 
 static int room_accept(MooCodeFrame *frame, MooObjectHash *env, MooArgs *args)
 {
-	args->m_result = new MooInteger(1);
+	args->m_result = new MooInteger((long int) 1);
 	return(0);
 }
